:tada: Se agrega soporte para cancelar solicitud
